[feat] 댓글 조회, 삭제 기능 수정
1. 자식 댓글이 존재하는지 여부를 확인하기 위해 쿼리를 요청하는 작업이 반복되기에 댓글 전체 리스트를 요청하여 필터링하는 방식으로 변경

2. 불필요한 쿼리 삭제 및 댓글 삭제 쿼리에서 userId 비교 구문 삭제

3. 서비스에 공통 부분 메소드로 분리

+++ src/main/java/me/sisyphusj/community/app/comment/service/CommentService.java
@@ -3,9 +3,9 @@ package me.sisyphusj.community.app.comment.service;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -54,9 +54,7 @@ public class CommentService {
 		List<CommentVO> newComments = new ArrayList<>();
 
 		// 댓글 ID, 자식 댓글 리스트를 매핑하는 해시 맵 생성
-		Map<Long, List<CommentVO>> childrenMap = comments.stream()
-			.filter(comment -> comment.getParentId() != null)
-			.collect(Collectors.groupingBy(CommentVO::getParentId));
+		Map<Long, List<CommentVO>> childrenMap = createChildrenMap(comments);
 
 		// 최상위 부모 댓글 리스트 생성
 		List<CommentVO> rootComments = comments.stream()
@@ -66,9 +64,7 @@ public class CommentService {
 		// 최상위 부모 댓글에서 자식 댓글 조회
 		for (CommentVO rootComment : rootComments) {
 			newComments.add(rootComment);
-			if (commentMapper.selectCountChildComment(rootComment.getCommentId()) > 0) {
-				addChildrenComments(rootComment, childrenMap, newComments);
-			}
+			addChildrenComments(rootComment, childrenMap, newComments);
 		}
 
 		// DTO로 변환하여 반환
@@ -94,16 +90,23 @@ public class CommentService {
 	 * 댓글 삭제
 	 */
 	@Transactional
-	public void removeComment(long commentId) {
+	public void removeComment(long postId, long commentId) {
 		// 삭제 대상 댓글 id 리스트
 		List<Long> deleteCommentIdList = new ArrayList<>();
 
-		// 작성자 확인 및 삭제하려는 댓글 조회
-		CommentVO commentVO = commentMapper.selectComment(SecurityUtil.getLoginUserId(), commentId)
-			.orElseThrow(CommentNotFoundException::new);
+		// 최신순으로 정렬된 댓글 리스트 가져오기
+		List<CommentVO> comments = commentMapper.selectCommentList(postId);
+
+		// 작성자 확인 및 삭제하려는 댓글이 존재하는지 확인
+		boolean isCommentExist = comments.stream()
+			.anyMatch(comment -> comment.getCommentId() == commentId && comment.getUserId() == SecurityUtil.getLoginUserId());
+
+		if (!isCommentExist) {
+			throw new CommentNotFoundException();
+		}
 
 		// 해당 댓글의 자식 댓글이 존재한다면 자식 댓글도 삭제
-		collectDeletableComments(commentVO.getCommentId(), deleteCommentIdList);
+		collectDeletableComments(commentId, comments, deleteCommentIdList);
 
 		// 작성자의 id 및 댓글 id 리스트를 통해 삭제 요청
 		commentMapper.deleteComment(SecurityUtil.getLoginUserId(), deleteCommentIdList);
@@ -124,26 +127,18 @@ public class CommentService {
 		Deque<CommentVO> stack = new ArrayDeque<>();
 
 		// 댓글 ID, 자식 댓글 리스트를 매핑하는 해시 맵 생성
-		Map<Long, List<CommentVO>> childrenMap = new HashMap<>();
+		Map<Long, List<CommentVO>> childrenMap = createChildrenMap(comments);
 
 		// 최상위 부모 댓글 스택 삽입, 자식 댓글은 해시 맵에 삽입
 		for (CommentVO comment : comments) {
 			if (comment.getParentId() == null) {
 				stack.push(comment);
-			} else {
-				childrenMap.computeIfAbsent(comment.getParentId(), k -> new ArrayList<>())
-					.add(comment);
 			}
 		}
 
 		while (!stack.isEmpty()) {
 			CommentVO comment = stack.pop();
 			newComments.add(comment);
-
-			// 자식 댓글이 존재하지 않으면 continue
-			if (commentMapper.selectCountChildComment(comment.getCommentId()) == 0) {
-				continue;
-			}
 
 			// 현재 댓글의 자식 댓글 리스트 조회
 			List<CommentVO> children = childrenMap.get(comment.getCommentId());
@@ -163,6 +158,15 @@ public class CommentService {
 	}
 
 	/**
+	 * 전체 댓글 리스트에서 댓글 ID, 자식 댓글 리스트를 매핑하는 해시 맵 반환
+	 */
+	private Map<Long, List<CommentVO>> createChildrenMap(List<CommentVO> comments) {
+		return comments.stream()
+			.filter(comment -> comment.getParentId() != null)
+			.collect(Collectors.groupingBy(CommentVO::getParentId));
+	}
+
+	/**
 	 * 현재 댓글 추가 및 자식 댓글을 재귀적으로 조회하여 추가
 	 *
 	 * @param currentComment 현재 댓글 객체
@@ -173,9 +177,9 @@ public class CommentService {
 		// 현재 댓글의 자식 댓글 리스트
 		List<CommentVO> children = childrenMap.get(currentComment.getCommentId());
 
-		for (CommentVO childComment : children) {
-			newComments.add(childComment);
-			if (commentMapper.selectCountChildComment(childComment.getCommentId()) > 0) {
+		if (children != null && !children.isEmpty()) {
+			for (CommentVO childComment : children) {
+				newComments.add(childComment);
 				addChildrenComments(childComment, childrenMap, newComments);
 			}
 		}
@@ -184,11 +188,12 @@ public class CommentService {
 	/**
 	 * 재귀적으로 자식 댓글의 ID를 조회하여 자신의 ID와 함께 deleteCommentIdList 에 저장
 	 */
-	private void collectDeletableComments(long commentId, List<Long> deleteCommentIdList) {
+	private void collectDeletableComments(long commentId, List<CommentVO> comments, List<Long> deleteCommentIdList) {
 		deleteCommentIdList.add(commentId);
 
 		// 자식 댓글 리스트 조회 후 각 댓글을 통해 메소드 호출
-		commentMapper.selectChildCommentIdList(commentId)
-			.forEach(childCommentId -> collectDeletableComments(childCommentId, deleteCommentIdList));
+		comments.stream()
+			.filter(comment -> Objects.equals(comment.getParentId(), commentId))
+			.forEach(comment -> collectDeletableComments(comment.getCommentId(), comments, deleteCommentIdList));
 	}
 }
